pyproject.toml
[tool.poetry]
name = "ecom-test-task"
version = "0.1.0"
description = "test task for eKom"
authors = ["Dimitry Rukhadze"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
tinydb = "^4.8.0"
djangorestframework = "^3.14.0"
environs = "^9.5.0"
phonenumbers = "^8.13.25"
python-dateutil = "^2.8.2"
pytest = "^7.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
